refactor(acp): package acp modules and extract app bootstrap
