+++ src/acp/mod.rs
@@ -1,0 +1,9 @@
+mod event_sink;
+mod runtime;
+
+pub use agenthub_acp::*;
+pub use event_sink::AgenthubAcpEventSink;
+pub(crate) use runtime::{
+    DEFAULT_ACTOR_CHANNEL, default_actor_cli_path, normalize_actor_cli_path,
+    normalize_actor_context,
+};
