+++ src/team/orchestrator.rs
@@ -6,8 +6,7 @@ use anyhow::Context;
 use async_trait::async_trait;
 use serde_json::Value;
 
-use crate::acp::AcpActorSkillContext;
-use crate::actor_runtime::{DEFAULT_ACTOR_CHANNEL, default_actor_cli_path};
+use crate::acp::{AcpActorSkillContext, DEFAULT_ACTOR_CHANNEL, default_actor_cli_path};
 use crate::agent::AgentManager;
 
 use super::{TeamManager, TeamRunRecord, TeamStepRecord, TeamStepStatus};
